backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
python-multipart==0.0.9
pydantic==2.8.2
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0